+++ apps/question-bank/src/test/java/gov/cdc/nbs/questionbank/page/content/question/UpdatePageCodedQuestionValuesetSteps.java
@@ -51,7 +51,7 @@ public class UpdatePageCodedQuestionValuesetSteps {
   public void page_coded_is_updated() throws Exception {
     response.active()
         .andExpect(status().isOk())
-        .andExpect(jsonPath("$.valueSet", equalTo(request.valueset())));
+        .andExpect(jsonPath("$.valueSet", equalTo((int)request.valueset())));
   }
 
 }
