+++ apps/question-bank/src/test/java/gov/cdc/nbs/questionbank/page/content/question/UpdatePageNumericQuestionSteps.java
@@ -102,7 +102,7 @@ public class UpdatePageNumericQuestionSteps {
   public void page_numeric_is_updated() throws Exception {
     response.active()
         .andExpect(status().isOk())
-        .andExpect(jsonPath("$.displayControl", equalTo(request.displayControl())))
+        .andExpect(jsonPath("$.displayControl", equalTo((int)request.displayControl())))
         .andExpect(jsonPath("$.fieldLength", equalTo(request.fieldLength())))
         .andExpect(jsonPath("$.defaultValue", equalTo(request.defaultValue().toString())))
         .andExpect(jsonPath("$.mask", equalTo(request.mask().toString())))
